metrics/QRelScore/requirements.txt
absl-py==0.11.0
aio-pika==6.4.1
aiohttp==3.7.4.post0
aiormq==3.3.1
alabaster==0.7.11
anaconda-client
appdirs==1.4.4
asgiref==3.4.1
asn1crypto==0.24.0
astor==0.8.1
astroid==2.0.4
astropy==3.0.4
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.9.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bce-python-sdk==0.8.53
beautifulsoup4==4.6.3
bert-score==0.3.8
bitarray==0.8.3
bkcharts==0.2
blaze
bleach==2.1.4
blessings==1.7
blis==0.4.1
bllipparser==2016.9.11
bokeh==0.13.0
boto==2.49.0
boto3==1.15.18
botocore==1.18.18
Bottleneck==1.2.1
brotlipy==0.7.0
cachetools==4.1.0
catalogue==2.0.6
certifi==2020.12.5
cffi
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cloudpickle==0.5.5
clyent
colorama==0.3.9
constantly==15.1.0
contextlib2==0.5.5
cPython==0.0.6
cryptography
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
cytoolz==0.9.0.1
dask==0.19.1
datasets==1.1.2
datashape
DAWG-Python==0.7.2
decorator==4.3.0
deeppavlov==0.14.0
defusedxml==0.5.0
Deprecated==1.2.13
dill==0.3.3
distlib==0.3.1
distributed==1.23.1
Django==3.2.9
django-cors-headers==3.10.0
docopt==0.6.2
docutils==0.14
dominate==2.5.1
einops==0.3.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastapi==0.47.1
fastcache==1.0.2
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.6
flatbuffers==1.12
fsspec==2021.6.1
funcsigs==1.0.2
future==0.18.2
fvcore==0.1.1.post20200716
gast==0.2.2
gdown==3.11.1
gensim==3.8.3
gevent==1.3.6
glob2==0.6
gmpy2==2.0.8
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpustat==0.6.0
graphviz==0.14
greenlet==0.4.15
grpcio==1.32.0
h11==0.9.0
h5py
heapdict==1.0.0
html5lib==1.0.1
httptools==0.1.1
hyperlink==18.0.0
identify==1.5.10
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==3.3.0
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
Keras-Applications
Keras-Preprocessing==1.1.2
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
littleutils==0.2.2
llvmlite==0.24.0
locket==0.2.0
loguru==0.5.3
lxml==4.2.5
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft
mkl-random
mkl-service
mock==4.0.2
more-itertools==4.3.0
moverscore==1.0.3
mpmath==1.0.0
msgpack==0.5.6
multidict==5.1.0
multipledispatch==0.6.0
multiprocess==0.70.11.1
murmurhash==1.0.2
nbconvert==5.4.0
nbformat==4.4.0
netaddr==0.8.0
networkx==2.1
nltk==3.4.5
nodeenv==1.5.0
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.18.0
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
objgraph==3.4.1
odo
olefile==0.46
opencv-python==4.2.0.32
openpyxl==2.5.6
opt-einsum==3.3.0
outdated==0.2.1
overrides==2.7.0
packaging==21.0
paddlepaddle==1.8.5
paddlepaddle-gpu==1.8.4.post107
pamqp==2.3.0
pandas==1.0.1
pandas-flavor==0.2.0
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib==1.0.1
pathlib2==2.3.2
pathy==0.6.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.6.0
pgl==1.1.0
pickleshare==0.7.4
Pillow==8.0.1
pingouin==0.3.11
pkginfo==1.4.2
plac==1.1.3
pluggy==0.7.1
ply==3.11
portalocker==2.0.0
pre-commit==2.9.3
prefetch-generator==1.0.1
preshed==3.0.2
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
py==1.6.0
py-readability-metrics==1.4.5
pyarrow==3.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycocotools==2.0.1
pycodestyle==2.6.0
pycorenlp==0.3.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.9.9
pycurl==7.43.0.2
pydantic==1.8.2
pyDeprecate==0.3.0
pydot==1.4.1
pyflakes==2.2.0
pyftpdlib==1.5.6
PyGithub==1.55
Pygments==2.2.0
PyJWT==2.3.0
pylint==2.1.1
pymongo==3.11.3
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.4.0
pyodbc==4.0.24
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrouge
PySocks==1.6.8
pyTelegramBotAPI==3.6.7
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
pytorch-lightning==1.3.8
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
rarfile==4.0
redis==2.10.6
redis-py-cluster==1.3.6
regex==2020.10.15
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.11.0
rsa==4.0
ruamel.yaml==0.15.100
rusenttokenize==0.0.5
s3transfer==0.3.3
sacremoses==0.0.35
scikit-image==0.18.1
scikit-learn==0.21.2
scipy==1.4.1
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
sentencepiece==0.1.91
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.15.0
smart-open==5.2.1
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
spacy==3.1.3
spacy-legacy==3.0.8
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
sqlparse==0.4.2
srsly==2.4.1
stanza==1.2
starlette==0.12.9
statsmodels==0.12.2
sympy==1.2
tables==3.4.4
tabulate==0.8.7
tblib==1.3.2
tensorboard==2.4.1
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tf-slim==1.1.0
Theano==1.0.5
thinc==8.0.10
thop==0.0.31.post2005241907
tifffile==2021.1.14
tokenizers
toml==0.10.2
toolz==0.9.0
torchaudio
torchfile
torchmetrics
torchsummary
torchtext
torchvision
tornado==5.1
tqdm==4.41.1
traitlets==4.3.2
transformers
Twisted==18.7.0
typer==0.4.0
typing==3.7.4.3
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.11
uvicorn==0.11.7
uvloop==0.15.2
virtualenv==20.2.2
visdom==0.1.8.9
visualdl==2.0.5
wasabi==0.8.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.4.1
wrapt==1.12.1
xarray==0.18.0
xdg==5.1.0
xgboost==1.1.1
xlrd==1.1.0
XlsxWriter==1.1.0
xlwt==1.3.0
xxhash==2.0.0
yacs==0.1.8
yarl==1.6.3
zict==0.1.3
zipp==3.4.0
zope.interface==4.5.0
